fix: compare MELBO activations at target layer

+++ tests/test_melbo.py
@@ -34,7 +34,8 @@ def small_transformer():
             return self
 
     mock_tokenizer.batch_encode_plus.return_value = MockEncoding(
-        input_ids=torch.randint(0, 12, (3, 8)), attention_mask=torch.ones(3, 8)
+        input_ids=torch.randint(0, 12, (3, 8)),
+        attention_mask=torch.ones(3, 8, dtype=torch.long),
     )
     model.tokenizer = mock_tokenizer
